Create fct vendas

+++ models/staging/erp_adventure_works/stg_erp__produtos.sql
@@ -1,7 +1,7 @@
 with
     source_product as (
         select 
-            cast (productid as numeric) as id_produto
+            cast (productid as int) as id_produto
             , cast (productsubcategoryid as string) as id_subcategoria_produto
             , cast (name as string) as nome_produto
             , cast (productnumber as string) as numero_produto
